Fixed some, but not all, compiler warnings: -Wconversion

+++ src/ttymidi.c
@@ -386,13 +386,13 @@ void write_midi_action_to_serial_port(snd_seq_t* seq_handle)
 
 void* read_midi_from_alsa(void* seq) 
 {
-	int npfd;
+	unsigned int npfd;
 	struct pollfd* pfd;
 	snd_seq_t* seq_handle;
 
 	seq_handle = seq;
 
-	npfd = snd_seq_poll_descriptors_count(seq_handle, POLLIN);
+	npfd = (unsigned int)snd_seq_poll_descriptors_count(seq_handle, POLLIN);
 	pfd = (struct pollfd*) alloca(npfd * sizeof(struct pollfd));
 	snd_seq_poll_descriptors(seq_handle, pfd, npfd, POLLIN);
